requirements.txt
pytest==6.2.5
pyyaml==6.0
requests==2.26.0
